+++ src/app/interfaces/verifiable-credential.ts
@@ -1,22 +1,52 @@
 export interface VerifiableCredential {
   '@context': string[];
   id: string;
-  type?: string[];
+  type?: string[];  
+  lifeCycleStatus: LifeCycleStatus; 
+  name?: string;
+  description?: string;
   issuer: Issuer;
-  issuanceDate: string;
   validFrom: string;
-  expirationDate: string;
   validUntil: string;
   credentialSubject: CredentialSubject;
-  available_formats?: string[];
-  status: CredentialStatus;
+  credentialStatus: CredentialStatus;
+  credentialEncoded?: string;
 }
 
 export interface Issuer {
   id: string;
+  organization?: string;
+  country?: string;
+  commonName?: string;
+  serialNumber?: string;
 }
 
-export interface CredentialSubject {
+export type CredentialSubject =
+  | EmployeeCredentialSubject
+  | LabelCredentialSubject
+  | MachineCredentialSubject;
+
+
+export interface LabelCredentialSubject {
+  id: string;
+  'gx:labelLevel': string;
+  'gx:engineVersion': string;
+  'gx:rulesVersion': string;
+  'gx:compliantCredentials': CompliantCredentials[];
+  'gx:validatedCriteria': string[];
+}
+
+export interface CompliantCredentials {
+  id: string;
+  type: string;
+  'gx:digestSRI': string;
+}
+
+export interface MachineCredentialSubject {
+  mandate: Mandate;
+}
+
+export interface EmployeeCredentialSubject  {
   mandate: Mandate;
 }
 
@@ -29,31 +59,44 @@ export interface Mandate {
 
 export interface Mandatee {
   id: string;
-  firstName: string;
-  lastName: string;
-  email: string;
-  nationality: string;
+  employeId?: string;
+  firstName?: string;
+  lastName?: string;
+  email?: string;
+  domain?: string;
+  ipAddress?: string;
 }
 
 export interface Mandator {
-  organizationIdentifier: string;
+  id: string;
   organization: string;
   commonName: string;
-  emailAddress: string;
   serialNumber: string;
   country: string;
 }
 
 export interface Power {
   id: string;
-  action: string | string[];
+  type: string;
   domain: string;
   function: string;
-  type: string;
+  action: string | string[];
 }
 
-export enum CredentialStatus {
-  VALID = 'VALID',
-  ISSUED = 'ISSUED',
-  REVOKED = 'REVOKED'
+export interface CredentialStatus {
+  id: string;
+  type: string;
+  statusPurpose: string;
+  statusListIndex: string;
+  statusListCredential: string;
 }
+
+export const LifeCycleStatuses = [
+  'VALID',
+  'ISSUED',
+  'REVOKED',
+  'EXPIRED',
+] as const;
+
+export type LifeCycleStatus = typeof LifeCycleStatuses[number];
+